limit initial columns to order count
There is no value in a column which has more items then
ordered. Limit count per length to order number. This might
also not be very helpful, but more then a useless column,
especially when low order counts are used.

Also use integer as data type to make clear we deal with
counts.